User authentication + User Listing relationship

+++ app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
         return redirect('/')->with('message', 'Listing created successfully.');
 
@@ -55,7 +57,10 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing){
         //dd($request->all());
-
+        // make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action.');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -76,8 +81,15 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized action.');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully.');
+    }
+
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
